+++ editor/db.inc.php
@@ -23,8 +23,7 @@ if (adminer()->homepage()) {
 		if ($name != "") {
 			echo '<tr><td>' . checkbox("tables[]", $table, in_array($table, (array) $_POST["tables"], true));
 			echo "<th><a href='" . h(ME) . 'select=' . urlencode($table) . "'>$name</a>";
-			$val = format_number($row["Rows"]);
-			echo "<td align='right'><a href='" . h(ME . "edit=") . urlencode($table) . "'>" . ($row["Engine"] == "InnoDB" && $val ? "~ $val" : $val) . "</a>";
+			echo "<td align='right'><a href='" . h(ME . "edit=") . urlencode($table) . "'>" . format_rows($row) . "</a>";
 		}
 	}